requirements.txt
aiogram>=3.0.0
aiohttp>=3.8.0
sqlalchemy>=2.0.0
alembic>=1.12.0
asyncpg>=0.29.0
redis>=5.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
apscheduler>=3.10.0
pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
ephem>=4.1.0
python-dateutil>=2.8.0
pytz>=2023.0
telethon>=1.28.0
easyocr>=1.7.0